pyproject.toml
[tool.poetry]
name = "dream-machine-hack"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lumaai = "^1.0.2"
ipykernel = "^6.29.5"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
openai = "^1.50.2"
python-dotenv = "^1.0.1"
streamlit = "^1.38.0"
boto3 = "^1.35.29"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
